Limit time spent waiting on stream reader threads
There may be cases where the stream reader can not finish even though
the subprocess has already terminated. For example the subprocess can
pass the stdout/stderr handles on to a daemon process which will keep
them open. In those cases we need to stop reading at some point to
continue the normal control flow.

Proposed solution: wait for 1 second for the stream reader to catch up,
and if the stream is still held open or the stream reader is still
running ignore any further output and continue.

Downside: this is essentially a memory leak as the stream reader thread
may still linger on. To limit the potential damage add a check for the
._closed property in the main stream reader loop and terminate there.

Might resolve #19.